add test/rename/etc
