+++ processing/src/main/java/org/apache/druid/query/rowsandcols/StorageAdapterRowsAndColumns.java
@@ -17,20 +17,18 @@
  * under the License.
  */
 
-package org.apache.druid.segment.join;
+package org.apache.druid.query.rowsandcols;
 
-import com.google.common.base.Preconditions;
 import org.apache.druid.frame.Frame;
 import org.apache.druid.frame.FrameType;
 import org.apache.druid.frame.allocation.ArenaMemoryAllocatorFactory;
 import org.apache.druid.frame.write.FrameWriter;
 import org.apache.druid.frame.write.FrameWriterFactory;
 import org.apache.druid.frame.write.FrameWriters;
+import org.apache.druid.java.util.common.ISE;
 import org.apache.druid.java.util.common.Intervals;
 import org.apache.druid.java.util.common.granularity.Granularities;
 import org.apache.druid.java.util.common.guava.Sequence;
-import org.apache.druid.query.rowsandcols.EmptyRowsAndColumns;
-import org.apache.druid.query.rowsandcols.RowsAndColumns;
 import org.apache.druid.query.rowsandcols.column.Column;
 import org.apache.druid.query.rowsandcols.concrete.FrameRowsAndColumns;
 import org.apache.druid.segment.CloseableShapeshifter;
@@ -48,12 +46,12 @@ import java.util.Collections;
 /**
  * Provides {@link RowsAndColumns} on top of a {@link StorageAdapter}.
  */
-public class StorageAdapterBasedRowsAndColumns implements CloseableShapeshifter, RowsAndColumns
+public class StorageAdapterRowsAndColumns implements CloseableShapeshifter, RowsAndColumns
 {
   private final StorageAdapter storageAdapter;
   private RowsAndColumns materialized;
 
-  public StorageAdapterBasedRowsAndColumns(StorageAdapter storageAdapter)
+  public StorageAdapterRowsAndColumns(StorageAdapter storageAdapter)
   {
     this.storageAdapter = storageAdapter;
   }
@@ -114,7 +112,11 @@ public class StorageAdapterBasedRowsAndColumns implements CloseableShapeshifter,
     RowSignature rowSignature = as.getRowSignature();
 
     FrameWriter writer = cursors.accumulate(null, (accumulated, in) -> {
-      Preconditions.checkNotNull("accumulated[%s] non-null, why did we get multiple cursors?", accumulated);
+      if (accumulated != null) {
+        // We should not get multiple cursors because we set the granularity to ALL.  So, this should never
+        // actually happen, but it doesn't hurt us to defensive here, so we test against it.
+        throw new ISE("accumulated[%s] non-null, why did we get multiple cursors?", accumulated);
+      }
 
       final ColumnSelectorFactory columnSelectorFactory = in.getColumnSelectorFactory();
